Move `Constants.cs` out of `Util/` folder
