+++ Constants.cs
@@ -1,0 +1,20 @@
+﻿namespace DSO
+{
+    static public class Constants
+    {
+        static public class Decompiler
+        {
+            public const string AUTHOR = "Elletra";
+            public const string VERSION = "0.1.0";
+            public const string EXTENSION = ".dso";
+
+            static public class GameVersions
+            {
+                public const uint BLV20 = 190;
+                public const uint BLV21 = 210;
+                public const uint TFD = 33;
+                public const uint TGE14 = 36;
+            }
+        }
+    }
+}
